DEV: now counting from orbitlist is finished and tested.
The counting can be done now for supercells with a constant memory footprint
for increasing size of supercell. The symmetry orbitlist is tested against the
distance based cluster equivalence and they seem to match for low cutoffs
(for larger cutoffs testing is more complicated).
Things that are not done is orbit sorting for orbitlist with crystal symmetry and
site symmetries/wyckoffs for the no crystal symmetry orbitlist. But that is out of
the scope for this branch.

+++ src/Structure.hpp
@@ -82,7 +82,7 @@ class Structure
         return _elements;
     }
 
-    int getElement(const size_t i) const
+    int getElement(const unsigned int i) const
     {
         if (i >= _elements.size())
         {
